\name{rankreg-internal}
\alias{RankRegEF}
\title{Internal rankreg functions}
\usage{
RankRegEF(y, d, x, beta, type = "Gehan")
}
\description{ Internal rankreg functions }
\details{
These are not intended to be called by the user.

\code{RankRegEF} computes the value of the
rank estimating function at given \code{beta}, and is used by
\code{RRV}.
}
\keyword{internal}

