\name{smallcell}
\alias{smallcell}
\title{Smallcell Lung Cancer Data}
\description{
There are 121 observations on 4 variables.
Arm is the indication of two treatments.
Entry is the age of the patient at entry.
Survival is the survival time and 
indicator is the censoring indicator (right censoring).
For more details please see the reference below.

Data are from 
Ying, Z., Jung, SH, and Wei, LJ (1995). 
Median regression analysis with censored data.
{\em Journal of the American Statistical Association},
{\bf 90}, 178-184.
}
\usage{data(smallcell)}
\format{A data frame containing 121 observations on 4 variables:
\tabular{rl}{
        [,1] \tab "arm"\cr
        [,2] \tab "entry"\cr
        [,3] \tab "survival"\cr
        [,4] \tab "indicator"}
}
\references{
{Ying, Z., Jung, SH, and Wei, LJ}
(1995). Median regression analysis with censored data.
{\em Journal of the American Statistical Association},
{\bf 90}, 178-184.
}
\keyword{datasets}
